feat: 🎸 create eposode
